1. Реализовать авторизация и аутентификацию

+++ middle_002/src/main/java/ru/inurgalimov/crud/presentation/UsersServlet.java
@@ -4,15 +4,19 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import ru.inurgalimov.crud.logic.Validate;
 import ru.inurgalimov.crud.logic.ValidateService;
+import ru.inurgalimov.crud.model.Role;
 import ru.inurgalimov.crud.model.User;
+import ru.inurgalimov.crud.utils.Utils;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Paths;
+import java.util.Objects;
 import java.util.UUID;
 
 public class UsersServlet extends HttpServlet {
@@ -39,15 +43,22 @@ public class UsersServlet extends HttpServlet {
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        req.setAttribute("users", ValidateService.getInstance().findAll());
-        req.getRequestDispatcher(PATH_FOR_SHOW_ALL_USERS).forward(req, resp);
+        HttpSession session = req.getSession();
+        Role userRole = (Role) session.getAttribute(Utils.KEY_FOR_GET_ROLE);
+        if (Objects.nonNull(userRole) && userRole.equals(Role.ADMINISTRATOR)) {
+            req.setAttribute(Utils.KEY_FOR_GET_USERS, ValidateService.getInstance().findAll());
+            req.getRequestDispatcher(PATH_FOR_SHOW_ALL_USERS).forward(req, resp);
+        }
     }
 
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         validate.delete(
                 new User(null, null, null, UUID.fromString(req.getParameter("id"))));
-        Files.deleteIfExists(Paths.get("images", req.getParameter("photoId")));
+        String photoId = req.getParameter("photoId");
+        if(Objects.nonNull(photoId) && !photoId.isEmpty()) {
+            Files.deleteIfExists(Paths.get("images", req.getParameter("photoId")));
+        }
         resp.sendRedirect(String.format("%s/list", req.getContextPath()));
     }
 }
